Using tmate/tmate-ssh-server container image directly
In the past, we have issue on generating keys for each container restart, this
will caused an issue if container crashed and restarted, the client has to have
a new `.tmate.conf`.

In this fix, we keep a generated SSH keys to a K8s secret, hence the pod with
that container can be restarted without generating new keys.

This also added a README.md to describe the general context of how K8s service
works using the tmate ssh server container image.

Last a new Makefile is added to streamline the service deployment.

Authored-by: Xin Zhang <xzhang@pivotal.io>